broke out the setup into a utils method
