+++ .eslintignore
@@ -1,3 +1,4 @@
 src/decls/**/*.js
 src/vendor/**/*.js
+src/test/utils.js
 *.test.js
